+++ Utilities/ScriptingLibrary/ScriptingExt.cs
@@ -8,7 +8,7 @@ namespace Sheepy.PhoenixPt.ScriptingLibrary {
    internal class ScriptingExt : ZyMod {
 
       internal static void RegisterAPI () {
-         Api( "api_add eval.cs", (ModnixAPI) API_Eval_CS );
+         Api( "api_add eval.js", (ModnixAPI) API_Eval_CS );
          //Api( "api_add zy.eval.cs", (ModnixAPI) API_Eval_CS );
       }
 
@@ -21,7 +21,7 @@ namespace Sheepy.PhoenixPt.ScriptingLibrary {
          return ScriptingEngine.Eval( mod, code );
       }
 
-      [ ConsoleCommand( Command = "Eval", Description = "(code) - Evaluate C# code, or enter C# shell if no code" ) ]
+      [ ConsoleCommand( Command = "Eval", Description = "(code) - Evaluate JavaScript code, or enter JavaScript shell if no code" ) ]
       public static void Console_Eval_CS ( IConsole console, string[] param ) {
          var code = string.Join( " ", param ?? Array.Empty<string>() ).Trim();
          if ( string.IsNullOrEmpty( code ) ) {